emoji: Data and Function to Work with Emojis

Contains data about emojis with relevant metadata, and functions to work with emojis when they are in strings.

See also

Author

Maintainer: Emil Hvitfeldt emilhhvitfeldt@gmail.com (ORCID)

Other contributors:

Hadley Wickham hadley@rstudio.com (Data parsing code from hadley/emo) [contributor]

Romain François (Data parsing code from hadley/emo) [contributor]