Full List of Emojis

This data set is the heart of the emoji package. It contains various information regarding all the avaiable emojis as of v15.0.

Usage

emojis

Format

tibble with 19 columns and nrow(emojis) rows

emoji

character representation of the emoji

name

name

group

group, e.g. "Smileys & People"

subgroup

sub group, e.g. "face-positive"

version

version where the emoji was introduced

points

Decimal Code Point(s)

nrunes

number of runes the emoji uses

runes

vector of unicode runes, i.e. hexadecimal representations prefixed with "U+"

qualified

Status of the emoji, can be one of 4 types; "component", "fully-qualified", "minimally-qualified", and "unqualified". See details for more.

vendor_* for apple ... windows

logical indicating if the given vendor supports the emoji

keywords

vector of keywords

keywords

vector of aliases

Source

Unicode® Full Emoji Charts v15.0

Unicode® Emoji Charts v15.0

Unicode® Emoji Ordering, v15.0

https://github.com/github/gemoji

https://github.com/muan/emojilib

Details

The levels of qualified have the following meaining

component: an Emoji_Component, excluding Regional_Indicators, ASCII, and non-Emoji.

fully-qualified: a fully-qualified emoji (see ED-18 in UTS #51), excluding Emoji_Component

minimally-qualified: a minimally-qualified emoji (see ED-18a in UTS #51)

unqualified: a unqualified emoji (See ED-19 in UTS #51)

See also